Exercises 2.1 - 2.5 completed
