--- part-1/courseinfo/src/Total.jsx
@@ -1,6 +1,0 @@
-const Total = ({ parts }) => {
-  const total = parts.map((part) => part.exercises).reduce((a, b) => a + b);
-  return <p>Number of exercises {total} </p>;
-};
-
-export default Total;
