--- part-1/courseinfo/src/main.jsx
@@ -1,4 +1,0 @@
-import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
-
-ReactDOM.createRoot(document.getElementById("root")).render(<App />);
