--- part-1/unicafe/src/App.jsx
@@ -1,27 +1,0 @@
-import { useState } from "react";
-import { Statistics } from "./Statistics";
-function App() {
-  const [good, setGood] = useState(0);
-  const [neutral, setNeutral] = useState(0);
-  const [bad, setBad] = useState(0);
-
-  return (
-    <div>
-      <h2>Give Feedback</h2>
-      <div>
-        <button onClick={() => setGood(good + 1)}>good</button>
-        <button onClick={() => setNeutral(neutral + 1)}>neutral</button>
-        <button onClick={() => setBad(bad + 1)}>bad</button>
-      </div>
-
-      <h2>Statistics</h2>
-      {good || bad || neutral ? (
-        <Statistics good={good} bad={bad} neutral={neutral} />
-      ) : (
-        <p>No feedback given</p>
-      )}
-    </div>
-  );
-}
-
-export default App;
