--- part-1/unicafe/src/StatisticLine.jsx
@@ -1,8 +1,0 @@
-export const StatisticLine = ({ text, value }) => {
-  return (
-    <tr>
-      <td>{text}</td>
-      <td>{value}</td>
-    </tr>
-  );
-};
