+++ part-2/courseinfo/package.json
