+++ part-2/courseinfo/src/components/Content.jsx
@@ -1,11 +1,14 @@
 import Part from "./Part";
+import Total from "./Total";
 
-const Content = ({ parts }) => {
+const Content = ({ parts, header }) => {
   return (
     <>
+      <h3>{header}</h3>
       {parts.map((part) => (
         <Part part={part.name} exercises={part.exercises} key={part.name} />
       ))}
+      <Total parts={parts} />
     </>
   );
 };
