+++ part-2/courseinfo/src/components/Part.jsx
