+++ part-2/courseinfo/src/components/Total.jsx
@@ -1,0 +1,6 @@
+const Total = ({ parts }) => {
+  const total = parts.reduce((total, part) => total + part.exercises, 0);
+  return <p style={{ fontWeight: "bold" }}>total of exercises {total} </p>;
+};
+
+export default Total;
