+++ part-2/courseinfo/src/main.jsx
@@ -1,4 +1,4 @@
 import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
+import App from "./components/App.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(<App />);
