+++ part-2/courseinfo/vite.config.js
